Simplify implementation of operation_state CPO
