+++ libs/core/execution_base/CMakeLists.txt
@@ -12,6 +12,7 @@ set(execution_base_headers
     hpx/execution_base/context_base.hpp
     hpx/execution_base/detail/spinlock_deadlock_detection.hpp
     hpx/execution_base/execution.hpp
+    hpx/execution_base/operation_state.hpp
     hpx/execution_base/receiver.hpp
     hpx/execution_base/register_locks.hpp
     hpx/execution_base/resource_base.hpp
