+++ libs/core/execution_base/include/hpx/execution_base/receiver.hpp
@@ -5,13 +5,12 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-// hpxinspect:nodeprecatedname:hpx::traits::is_callable
-
 #pragma once
 
 #include <hpx/config/constexpr.hpp>
 #include <hpx/functional/tag_fallback_invoke.hpp>
 #include <hpx/functional/tag_invoke.hpp>
+#include <hpx/functional/traits/is_invocable.hpp>
 
 #include <exception>
 #include <type_traits>
