+++ libs/core/execution_base/tests/unit/CMakeLists.txt
@@ -1,10 +1,11 @@
 # Copyright (c) 2019 Thomas Heller
+# Copyright (c) 2020 Hartmut Kaiser
 #
 # SPDX-License-Identifier: BSL-1.0
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-set(tests basic_receiver basic_operation_state execution_context)
+set(tests basic_operation_state basic_receiver execution_context)
 
 foreach(test ${tests})
   set(sources ${test}.cpp)
